AI_Fitness_and_Wellness_Coach/requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.1
firebase-admin>=6.5.0
google-generativeai>=0.7.2
google-auth>=2.29.0
requests>=2.32.0
gunicorn>=21.2.0